Feat: Add details.jsp, AdDetailsServlet.java logic, index.jsp logic

+++ src/main/java/com/codeup/adlister/dao/Ads.java
@@ -11,5 +11,5 @@ public interface Ads {
     Long insert(Ad ad);
     
 
-    Ad findById(int id);
+    Ad findById(long id);
 }
